murmur: fix hash to match cassandra sign bug (#1035)
Cassandra has a bug in the murmur hash which causes it to use signed
shifts. Fix this and refactor the code to make it easier to reason
about. All the extracted functions are inlined so there is no
performance change.

+++ token.go
@@ -39,7 +39,7 @@ func (p murmur3Partitioner) Name() string {
 
 func (p murmur3Partitioner) Hash(partitionKey []byte) token {
 	h1 := murmur.Murmur3H1(partitionKey)
-	return murmur3Token(int64(h1))
+	return murmur3Token(h1)
 }
 
 // murmur3 little-endian, 128-bit hash, but returns only h1
